fix(runtime,evaluator): DWScript float precision and two's-complement hex
Two output-formatting gaps surfaced by FunctionsByteBuffer/floats and
FunctionsByteBuffer/integers:

- Float rendering used Go's shortest round-trip form, so 3.141592025756836
  printed 16 significant digits where DWScript's FloatToStr prints 15
  (3.14159202575684), and 15-digit integral values such as -845680067215360
  came out in exponent form. Both now use %.15g.
- Integer.ToHexString rendered a negative value with a minus sign ("-80")
  instead of its 64-bit two's-complement pattern ("FFFFFFFFFFFFFF80").

Both are corpus-wide fixes rather than ByteBuffer-specific ones; measured
against the full fixture suite they move only FunctionsByteBuffer and regress
no category.

+++ internal/interp/evaluator/integer_helpers.go
@@ -85,8 +85,10 @@ func (e *Evaluator) evalIntegerToHexString(selfValue Value, args []Value, node a
 		digits = 0
 	}
 
-	// Format as uppercase hex
-	hexStr := fmt.Sprintf("%X", intVal.Value)
+	// Format as uppercase hex. A negative value is rendered as its 64-bit
+	// two's-complement pattern, as DWScript does: (-128).ToHexString(1) is
+	// "FFFFFFFFFFFFFF80", not "-80".
+	hexStr := fmt.Sprintf("%X", uint64(intVal.Value))
 
 	// Pad with leading zeros if needed
 	if len(hexStr) < digits {
